+++ include/ccf/pal/attestation.h
@@ -11,9 +11,6 @@
 #include "ccf/pal/measurement.h"
 #include "ccf/pal/snp_ioctl.h"
 
-// TODO: Public->private
-#include "ds/files.h"
-
 #include <fcntl.h>
 #include <functional>
 
@@ -222,72 +219,6 @@ namespace ccf::pal
       }
     }
   }
-
-  static std::string virtual_attestation_path(const std::string& suffix)
-  {
-    return fmt::format("ccf_virtual_attestation.{}.{}", ::getpid(), suffix);
-  };
-
-  static void emit_virtual_measurement(
-    const std::string& package_path, const std::string& security_policy)
-  {
-    auto package = files::slurp(package_path);
-
-    auto package_hash = ccf::crypto::Sha256Hash(package);
-
-    auto j = nlohmann::json::object();
-    j["measurement"] = package_hash.hex_str();
-    j["security_policy"] = security_policy;
-
-    files::dump(j.dump(2), virtual_attestation_path("measurement"));
-  }
-
-#if defined(PLATFORM_VIRTUAL)
-
-  static void generate_quote(
-    PlatformAttestationReportData& report_data,
-    RetrieveEndorsementCallback endorsement_cb,
-    const snp::EndorsementsServers& endorsements_servers = {})
-  {
-    auto quote = files::slurp_json(virtual_attestation_path("measurement"));
-    quote["report_data"] = ccf::crypto::b64_from_raw(report_data.data);
-
-    files::dump(quote.dump(2), virtual_attestation_path("attestation"));
-
-    auto dumped_quote = quote.dump();
-    std::vector<uint8_t> quote_vec(dumped_quote.begin(), dumped_quote.end());
-
-    endorsement_cb(
-      {.format = QuoteFormat::insecure_virtual,
-       .quote = quote_vec,
-       .endorsements = {},
-       .uvm_endorsements = {},
-       .endorsed_tcb = {}},
-      {});
-  }
-
-#elif defined(PLATFORM_SNP)
-
-  static void generate_quote(
-    PlatformAttestationReportData& report_data,
-    RetrieveEndorsementCallback endorsement_cb,
-    const snp::EndorsementsServers& endorsements_servers = {})
-  {
-    QuoteInfo node_quote_info = {};
-    node_quote_info.format = QuoteFormat::amd_sev_snp_v1;
-    auto attestation = snp::get_attestation(report_data);
-
-    node_quote_info.quote = attestation->get_raw();
-
-    if (endorsement_cb != nullptr)
-    {
-      endorsement_cb(
-        node_quote_info,
-        snp::make_endorsement_endpoint_configuration(
-          attestation->get(), endorsements_servers));
-    }
-  }
-#endif
 
 #if !defined(INSIDE_ENCLAVE) || defined(VIRTUAL_ENCLAVE)
